+++ intelmq/lib/message.py
@@ -143,6 +143,14 @@ class Message(dict):
                 else:
                     raise
 
+    def __delitem__(self, item):
+        print('delitem', item)
+        if item == 'extra':
+            for key in [key for key in self.keys() if key.startswith('extra.')]:
+                del self[key]
+            return
+        return super().__delitem__(item)
+
     def get(self, key, default=None):
         try:
             return self[key]
@@ -252,6 +260,9 @@ class Message(dict):
         class_name, subitem = self.__get_type_config(key)
         if class_name and class_name['type'] == 'JSONDict' and not subitem:
             # for backwards compatibility allow setting the extra field as string
+            if overwrite and key in self:
+                print(overwrite, key in self, self[key])
+                del self[key]
             for extrakey, extravalue in json.loads(value).items():
                 if hasattr(extravalue, '__len__'):
                     if not len(extravalue):  # ignore empty values